+++ No_ssu_backend/settings.py
@@ -54,7 +54,6 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'marker',
 
     # My apps
     'api',
