--- marker/views.py
@@ -1,6 +1,0 @@
-from django.shortcuts import render
-from .models import Marker
-
-def home(request):
-    markers = Marker.objects
-    return render(request,'')
